requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
channels-yroom==0.0.6
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.5
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-filter==23.3
django-nine==0.2.7
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.7
packaging==23.1
psycopg==3.1.10
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==1.26.16
yroom==0.0.10
zope.interface==6.0
